SCRUM-211 refactor: reorganize imports and update type definitions for payment flow and account modal
