+++ src/features/subscriptions/model/payments.types.ts
@@ -4,3 +4,15 @@ export type PaymentsSortState = {
   key: PaymentsSortBy | null
   direction: PaymentsSortDirection | null
 }
+
+export type PaymentFlowStatus = 'idle' | 'success' | 'failure'
+
+export type PaymentFlowState = {
+  flowStatus: PaymentFlowStatus
+  flowErrorCode: null | string
+  isStarting: boolean
+  startPayment: () => Promise<void>
+  resetFlow: () => void
+}
+
+export type AccountModal = null | 'auto' | 'success' | 'failure'
